Add sui_utils crate, move trace_utils to it
