+++ sui/Cargo.toml
@@ -32,7 +32,6 @@ serde-value = "0.7.0"
 serde-name = "0.2.0"
 dirs = "4.0.0"
 clap = { version = "3.1.8", features = ["derive"] }
-telemetry_subscribers = { git = "https://github.com/MystenLabs/mysten-infra", rev = "a45af6dc28aa12c8cc13521d118c24aadd4c6adf" }
 
 bcs = "0.1.3"
 sui_core = { path = "../sui_core" }
@@ -40,6 +39,7 @@ sui-adapter = { path = "../sui_programmability/adapter" }
 sui-framework = { path = "../sui_programmability/framework" }
 sui-network = { path = "../network_utils" }
 sui-types = { path = "../sui_types" }
+sui-utils = { path = "../sui_utils" }
 sui-verifier = { path = "../sui_programmability/verifier" }
 
 rustyline = "9.1.2"
