+++ sui/src/lib.rs
@@ -16,7 +16,6 @@ pub mod rest_gateway;
 pub mod shell;
 pub mod sui_commands;
 pub mod sui_json;
-pub mod trace_utils;
 pub mod wallet_commands;
 
 const SUI_DIR: &str = ".sui";
